% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{helpers}
\alias{rvattr}
\alias{rvtype}
\alias{tabletype}
\alias{is_rvtable}
\alias{is_sample}
\alias{is_distribution}
\alias{is_discrete}
\alias{is_continuous}
\alias{is_density}
\alias{valcol}
\alias{probcol}
\alias{idcols}
\alias{get_density_args}
\alias{get_sample_args}
\alias{set_density_args}
\alias{set_sample_args}
\title{rvtable helpers and attributes}
\usage{
rvattr(x, id, all = FALSE)

rvtype(x)

tabletype(x)

is_rvtable(x)

is_sample(x)

is_distribution(x)

is_discrete(x)

is_continuous(x)

is_density(x)

valcol(x)

probcol(x)

idcols(x)

get_density_args(x)

get_sample_args(x)

set_density_args(x, density.args)

set_sample_args(x, sample.args)
}
\arguments{
\item{x}{rvtable.}

\item{id}{character, rvtable attribute(s). If not provided, all available rvtable attributes.}

\item{all}{logical, ignore `id` and return all attributes,
including those not specific to the `rvtable` class. Defaults to \code{FALSE}.}

\item{density.args}{set rvtable \code{density} arguments. See details.}

\item{sample.args}{set rvtable sampling arguments. See details.}
}
\value{
information about various rvtable attributes. See details.
}
\description{
Helper functions for the rvtable class.
}
\details{
These helper functions provide information about `rvtable` class objects and their attributes.
They get and in some cases set rvtable attributes as well as perform various logical checks.
\code{rvattr} returns a named list of attributes of \code{x}. \code{is*} functions return \code{TRUE} or \code{FALSE}.
\code{rvtype}, \code{tabletype}, \code{valcol} and \code{probcol} return individual attributes of \code{x} (character).

\code{get*} functions for density and sample argument lists return the lists from the attributes of \code{x}
and the corresponding \code{set*} functions return \code{x} with the new \code{denisty.args} and \code{sample.args} attributes.
Setting density or sample argument lists is usually done when passing arguments to a function, e.g., \code{rvtable}
or \code{sample_rvtable}. However, it can sometimes be useful to set these directly or update them at a desired step
in a rvtable processing pipeline. Note that when defaults are not explicitly provided (i.e., leaving as \code{list()} or \code{NULL}),
the atrributes of an rvtable will still display all sampling default arguments and the most important \code{density}
default arguments for clarity.
}
\examples{
x <- rvtable(1:10)

is_rvtable(x)
is_sample(x)
is_distribution(x)
is_discrete(x)
is_continuous(x)
is_density(x)

rvattr(x)
rvattr(x, id=c("rvtype", "tabletype"))
rvattr(x, all = TRUE)

rvtype(x)
tabletype(x)
valcol(x)
probcol(x)

get_density_args(x)
get_sample_args(x)
x <- set_density_args(x, list(n = 1000, adjust = 0.1))
x <- set_sample_args(x, list(n = 100))
get_density_args(x)
get_sample_args(x)
}
