% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_pmf.R
\name{inverse_pmf}
\alias{inverse_pmf}
\title{Inverse Probability Mass Function}
\usage{
inverse_pmf(x, val.range, var.new, sample.args = list())
}
\arguments{
\item{x}{an rvtable.}

\item{val.range}{range of values of the continuous random variable in the rvtable.}

\item{var.new}{the categorical variable for which to compute the pmf given \code{val.range}.}

\item{sample.args}{optional arguments used when sampling.}
}
\value{
an rvtable.
}
\description{
Compute the probability mass function for a specified categorical variable conditional on an interval of the continuous random variable's values.
}
\details{
This function computes the pmf of a categorical variable, providing probabilities corresponding to the levels of the variable, given a specific range of values of the continuous random variable in the rvtable.
At this time the rvtable type must be continuous.
}
\examples{
library(data.table)
library(dplyr)
x <- data.table(
  id1=rep(LETTERS[1:5], each=4),
  id2=factor(c("low", "high")),
  id3=rep(1:2, each=2),
  Val=rep(1:10, each=20), Prob=rep(sqrt(1:10), each=20)) \%>\% rvtable
y1 <- inverse_pmf(x, c(5, 8), "id1", sample.args=list(n=5))
y1
x <- filter(x, id2=="low" & id3==1) \%>\% select(-id2, -id3) \%>\% rvtable
y2 <- inverse_pmf(x, c(5,8), "id1", sample.args=list(n=5))
y2
}

