% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvtable.R
\name{is.rvtable}
\alias{is.rvtable}
\title{Check For rvtable Class}
\usage{
is.rvtable(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
\code{TRUE} if \code{x} is an rvtable class object, otherwise \code{FALSE}.
}
\description{
Check if an object has class rvtable.
}
\examples{
is.rvtable("a")
is.rvtable(rvtable(1:10))
}
