% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{marginalize}
\alias{marginalize}
\title{Marginal Distributions}
\usage{
marginalize(x, margin, density.args, sample.args)
}
\arguments{
\item{x}{an rvtable.}

\item{margin}{variable(s) in rvtable to marginalize over.}

\item{density.args}{optional arguments passed to \code{density}. If supplied, overrides the \code{density.args} attribute of \code{x}.}

\item{sample.args}{optional arguments used when sampling. If supplied, overrides the \code{sample.args} attribute of \code{x}.}
}
\value{
an rvtable.
}
\description{
Obtain a marginal distribution of a random variable in an rvtable.
}
\details{
Grouping variables are ignored when marginalizing the distribution of a random variable over explicit categorical variables.
Unlike \code{merge_rvtable}, \code{marginalize} will not merge ungrouped ID variables.
It will only marginalize over the ID variables specified in \code{margin}.
It will also account for unequal weighting of an ID variable's levels, taken from the \code{weights} attribute of an rvtable,
whereas \code{merge_rvtable} is a simpler function that ignores weights.
}
\examples{
\dontrun{
library(dplyr)
x <- data.frame(
  id1=rep(LETTERS[1:5], each=4),
  id2=factor(c("low", "high")),
  id3=rep(1:2, each=2),
  Val=rep(1:10, each=20), Prob=rep(sqrt(1:10), each=20)) \%>\% rvtable
marginalize(x, c("id1", "id2"))

get_levels(x, "id1")
get_weights(x, "id1")
marginalize(x, "id1")
wts <- data.frame(levels=LETTERS[1:5], weights=c(1, 1.5, 2, 4, 1))
x <- set_weights(x, "id1", wts)
get_weights(x, "id1")
marginalize(x, "id1")
}
}
