% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvtable.R
\name{rvtable}
\alias{rvtable}
\title{Random Variable Table}
\usage{
rvtable(x, y = NULL, Val = "Val", Prob = "Prob", discrete = FALSE,
  density.args = list())
}
\arguments{
\item{x}{a numeric vector, data frame, or data table.}

\item{y}{an optional vector of probabilities associated with \code{x} when \code{x} is a numeric vector with no similar probabilities vector attribute.}

\item{Val}{the column name of \code{x} referring to random variable values when \code{x} is a data frame or data table.}

\item{Prob}{the column name of \code{x} referring to random variable values when \code{x} is a data frame or data table.}

\item{discrete}{whether the random variable is discrete.}

\item{density.args}{optional arguments passed to \code{density}.}
}
\value{
an S3 object of class \code{rvtable}
}
\description{
Class constructor for \code{rvtable} objects.
}
\details{
These are long format data tables containing a Val and Prob column describing the distribution of a random variable.
Any other columns are ID columns and should be categorical variables. The random variable described by Val and Prob may be discrete or continuous.
When discrete, probabilities are true probabilities.
When continuous, Val and Prob are based on \code{x} and \code{y} output from \code{density} and describe a distribution curve,
and therefore values in Prob may be greater than one and may not sum to one.
Val is typically numeric but may be character when discrete such as when an rvtable object is returned from \code{inverse_pmf}.
All rvtable objects are either distribution-based or sample-based.
This primary constructor only constructs distribution- or sample-based rvtable objects, with the attribute \code{tabletype="distribution"} or \code{tabletype="sample"}.
Sampling on an rvtable can generate a sample-based rvtable, with the attribute \code{tabletype="sample"}.
Every rvtable object also has a variable type attribute, \code{rvtype}, which is either "discrete" or "continuous".
}
\examples{
# basic samples from continuous and discrete RVs
x <- rnorm(100)
rvtable(x)
rvtable(x, density.args=list(n=50, adjust=2))
rvtable(x, discrete=TRUE) # incorrect: x is a continuous RV
x <- sample(1:10, size=30, replace=TRUE, prob=sqrt(10:1))
rvtable(x, discrete=TRUE) # discrete=T only needed if ambiguous
x <- 1:5
probs <- c(0.1, 0.2, 0.3, 0.15, 0.25)
rvtable(x, y=probs) # discrete inferred from y
attr(x, "probabilities") <- probs
rvtable(x) # discrete inferred from attributes

# an existing data frame or data table
x <- data.frame(Val=1:10, Prob=0.1)
rvtable(x)
library(data.table)
x <- data.table(id=rep(LETTERS[1:2], each=10), v1=rep(1:10, 2), p1=c(rep(0.1, 10), sqrt(1:10)))
rvtable(x, Val="v1", Prob="p1")
}
