% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights}
\alias{weights}
\alias{get_weights}
\alias{set_weights}
\title{ID variable level weights}
\usage{
get_weights(x, id)

set_weights(x, id, weights)
}
\arguments{
\item{x}{rvtable.}

\item{id}{character, ID column name in \code{x}.}

\item{weights}{list, data frame, or \code{NULL}. See details.}
}
\value{
a list for \code{get_weights}.
}
\description{
Get or set ID variable level weights.
}
\details{
Weights are important for marginalizing distributions, when collapsing an rvtable over levels of an ID
variable must account for known unequal weighting of levels. Weights are strictly used by \code{marginalize}.

\code{get_weights} get a list of 2-column data frames of levels and associated weights for ID variables in an rvtable.
\code{set_weights} sets weights for levels of ID variables in an rvtable.
Weights for levels of ID variables in an rvtable are assumed to be equal unless explicitly set otherwise.
They are usually set via \code{set_weights} but can also be set as part of a call to the \code{rvtable} constructor.

If \code{weights=NULL}, \code{id} is ignored and all ID variables' levels are set to equal weighting.
If \code{id} refers to a single ID variable in \code{x}, then \code{weights} can be a valid weights data frame.
In general, \code{weights} is a list of data frames corresponding to the variables specified \code{id}.
If \code{weights} is a named list, the named map to \code{id}. If unnamed, they must be provided in the same order as \code{id}.
Equal weights are set for an individual ID variable's levels
when a list element in \code{weights} correpsonding to an \code{id} value is \code{NULL}.
}
\examples{
x <- rvtable(data.frame(
  id1=rep(LETTERS[1:5], each=4),
  id2=factor(c("low", "high")),
  id3=rep(1:2, each=2),
  Val=rep(1:10, each=20), Prob=rep(sqrt(1:10), each=20)))
get_weights(x)
get_weights(x, "id1")
x <- set_weights(x, "id1", data.frame(levels=LETTERS[1:5], weights=1:5))
get_weights(x, c("id1", "id3"))
wts <- data.frame(levels=1:2, weights=c(0.3, 0.7))
x <- set_weights(x, c("id1", "id3"), list(id1=NULL, id3=wts))
get_weights(x, c("id1", "id3"))
wts <- data.frame(levels=levels(x$id2), weights=c(2, 1))
x <- set_weights(x, c("id1", "id2"), list(NULL, wts))
get_weights(x)
}
