% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{cycle_rvtable}
\alias{cycle_rvtable}
\title{Repeated Resampling Utility}
\usage{
cycle_rvtable(x, n, start = NULL, density.args = list(),
  sample.args = list())
}
\arguments{
\item{x}{an rvtable in distribution form only.}

\item{n}{total number of iteratively estimated distributions to return, including the original from \code{x}.}

\item{start}{usually \code{NULL} when first called and on subsequent recursive calls an integer representing which cycle to begin from in the updated rvtable.}

\item{density.args}{optional arguments passed to \code{density}.}

\item{sample.args}{optional arguments used when sampling.}
}
\value{
an rvtable.
}
\description{
Repeat cycles of bootstrap resampling followed by density re-estimation for testing purposes.
}
\details{
This function repeats cycles of bootstrap resampling from a distibution followed by distribution re-estimation \code{n - 1} times.
\code{x} must be a distribution-form rvtable and not a sample-form rvtable.
}
\examples{
library(data.table)
library(dplyr)
x <- data.table(
  id1=rep(LETTERS[1:5], each=4),
  id2=factor(c("low", "high")),
  id3=rep(1:2, each=2),
  Val=rep(1:10, each=20), Prob=rep(sqrt(1:10), each=20)) \%>\% rvtable
cycle_rvtable(x, 2)
x \%>\% group_by(id1, id2) \%>\% cycle_rvtable(3)
}

