% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{marginalize}
\alias{marginalize}
\title{Marginal Distribution rvtable}
\usage{
marginalize(x, margin, weights = NULL, density.args = list(),
  sample.args = list())
}
\arguments{
\item{x}{an rvtable.}

\item{margin}{variable(s) in rvtable to marginalize over.}

\item{weights}{relative weights for unique values or levels of a single \code{margin} variable.}

\item{density.args}{optional arguments passed to \code{density}.}

\item{sample.args}{optional arguments used when sampling.}
}
\value{
an rvtable.
}
\description{
Obtain a marginal distribution of a random variable in an rvtable.
}
\details{
Grouping variables are ignoed when marginalizing the distribution of a random variable over explicit categorical variables. \code{margin} must be explicit.
\code{weights} only applies in the clear case of marginalizing over a single categorical variable.
Marginalizing over multiple variables in a single call to \code{marginalize} is only available assuming equal weights for all values of those variables.
When using weights, \code{marginalize} must be called on one variable at a time.
Call \code{get_levels} on an rvtable first to ensure weights are passed in the correct order.
}
\examples{
library(data.table)
library(dplyr)
x <- data.table(
  id1=rep(LETTERS[1:5], each=4),
  id2=factor(c("low", "high")),
  id3=rep(1:2, each=2),
  Val=rep(1:10, each=20), Prob=rep(sqrt(1:10), each=20)) \%>\% rvtable
marginalize(x, c("id1", "id2"))
get_levels(x, "id1")
marginalize(x, "id1", weights=c(1, 1.5, 2, 4, 1))
}

