% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{merge_rvtable}
\alias{merge_rvtable}
\title{Merge RV Table Conditional Distributions}
\usage{
merge_rvtable(x, density.args = list(), sample.args = list())
}
\arguments{
\item{x}{an rvtable.}

\item{density.args}{optional arguments passed to \code{density}.}

\item{sample.args}{optional arguments used when sampling.}
}
\value{
an rvtable.
}
\description{
Merge conditional distributions of a random variable in an rvtable over levels of ungrouped categorical variables.
}
\details{
Distributions are merged using a cycle of bootstrap resampling followed by density re-estimation.
Merging relies simply on what \code{x} is grouped by.
It assumes equal weights for all levels of all grouping variables over which a random variable is being merged.
This is why it is called merging rather than marginalizing.
This function is also used by \code{marginalize} and can properly obtain a marginal distribution when passed an rvtable including a \code{weights} column, which \code{marginalize} provides.
If calling this function directly, it is important to know that it is generally intended for simple merging.
Use \code{marginalize} to ensure proper marginal distributions are obtained since it can be used with or without non-constant weights.
}
\examples{
\dontrun{
library(data.table)
library(dplyr)
x <- data.table(
  id1=rep(LETTERS[1:5], each=4),
  id2=factor(c("low", "high")),
  id3=rep(1:2, each=2),
  Val=rep(1:10, each=20), Prob=rep(sqrt(1:10), each=20)) \%>\% rvtable
merge_rvtable(x)
x \%>\% group_by(id1) \%>\% merge_rvtable
x \%>\% group_by(id1, id2) \%>\% merge_rvtable
}
}
