% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvdplyr.R
\name{rvdplyr}
\alias{rvdplyr}
\alias{arrange.rvtable}
\alias{filter.rvtable}
\alias{slice.rvtable}
\alias{mutate.rvtable}
\alias{summarise.rvtable}
\alias{summarize.rvtable}
\alias{distinct.rvtable}
\title{Preserving rvtable class and attributes with dplyr}
\usage{
\method{arrange}{rvtable}(.data, ...)

\method{filter}{rvtable}(.data, ...)

\method{slice}{rvtable}(.data, ...)

\method{mutate}{rvtable}(.data, ...)

\method{summarise}{rvtable}(.data, ...)

\method{summarize}{rvtable}(.data, ...)

\method{distinct}{rvtable}(.data, ...)
}
\arguments{
\item{.data}{rvtable.}

\item{...}{additional arguments.}
}
\value{
an rvtable.
}
\description{
Methods for the rvtable class are provided for a number of common generic \code{dplyr} methods.
}
\details{
Functions in the \code{dplyr} package tend to strip custom classes and attributes from data frames.
These rvtable methods preserve the \code{rvtable} class as the primary class and preserve all rvtable-specific attributes.
}
